pyproject.toml
[project]
name = "coloring-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.8.2",
    "django>=5.2.8",
    "flask>=3.1.2",
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "zipfile36>=0.1.3",
]
